requirements.txt
torch==2.4.1+cu118 --index-url https://download.pytorch.org/whl/cu118
torchvision==0.19.1+cu118 --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.4.1+cu118 --index-url https://download.pytorch.org/whl/cu118
numpy>=1.24,<1.27
pandas>=2.0,<2.1
matplotlib>=3.7,<3.8
seaborn>=0.13
scikit-image==0.21.*
scikit-learn==1.3.*
nibabel>=5.2
SimpleITK>=2.4
nilearn>=0.10
tensorboard>=2.14
wandb>=0.22
pillow>=10
tqdm>=4.66
pyyaml>=6
scipy>=1.10,<1.11
statsmodels>=0.14
opencv-python-headless>=4.10 ; platform_system!="Windows"  # optional if you prefer cv2 on non-Windows